+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.3.0] - 2023-06-04
 * Add class `BWidgets::Activatable` and implement in `BWidgets::Widget`
+* Add class `BWidgets::Navigatable` and implement in `BWidgets::Box`
 
 
 ## [1.2.2] - 2023-06-02